requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
crispy-tailwind==0.5.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-autoslug==1.9.8
django-cities==0.6
django-cities-light==3.8.1
django-crispy-forms==1.12.0
django-model-utils==4.1.1
idna==3.2
oauthlib==3.1.1
Pillow==8.3.1
progressbar2==3.53.1
pycparser==2.20
PyJWT==2.1.0
python-utils==2.5.6
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
tqdm==4.62.0
Unidecode==1.2.0
urllib3==1.26.6
